remove "Limitations"
1. Positional parameters
Since positional params have to be mapped to either individual properties or a single array property, it is possible to refer to them as named arguments.
2. Named arguments
Calling named arguments a limitation feels counter to the fact that they allow passing HTML attributes.
3. Capitalization
Again, it seems strange to call the design of the first letter needing to be capital a limitation.

These should more likely be documented as part of the design of angle bracket invocation syntax.