Fix asynchronous action component test example
- The build complains that there's no close to `<Input>`.
- The linter complains about single quotes in handlebars.
- The linter complains about the {{#each}} indentation.

Incidentally I cannot get this test to pass in a fresh Ember 3.11.1 app. I tried running the tests in Chrome, CLI, Firefox, Safari. No matter what I do it appears that the `dispatchEvent` does not cause the `fetchResults` function to be called. I can't tell if this is because `dispatchEvent` won't trigger that or if `@fetchResults={{this.fetchResults}}` is not attaching the handler properly or some other problem.