Try a different CSS approach
Using `position: relative; left: -1ch;` had a very mild side effect
that when you make text selection, the browser seems a tad confused
bout how to draw the highlight color.

It doesn't affect the selection/copy/paste functionality, as far as
I can tell, just looks a bit odd visually. It occured to me there
may be another way to pull the text to the left without that problem
so here it is.

It seems to work the same, but does not have the selection rendering
issue.