small clarification on implementing Ember.onerror
In modern apps it's required to import ember now, and it's not an import new users will be used to, so it seems logical to add it here. I also suggested to add it in app/app.js, but if there is a better place let me know and I'll update it :).