upgrading / tracked properties / backwards-compatibility: use the locals https://guides.emberjs.com/release/upgrading/current-edition/tracked-properties/#toc_backwards-compatibility
The example describes using get to read a property that isn't @tracked,
saying that they will autotrack as long as you use get.
The example below it says that if you don't use get it won't autotrack.

The trouble is, the example loads the properties with get into locals
then doesn't use them. That confused me: was the aim to side-effect the
property, or was it to return the values with tracking attached to them,
but not tracking the original property if you then access it directly?

This commit assumes the latter, and changes the example to use the
variables, not reload the property directly.