Make backburner docs more explicit
I'm not entirely sure that this is the right place to put it, but current wording is pretty vague.

Do I need to add it to my application code? To my tests? In `beforeEach` of my tests?

Also current version makes an assumption that the user knows that they need to import `Ember`.