controllers: fix broken link to API docs
The link to the Controller API docs is broken, so when anyone clicks it
they'll be met with a 404 error. Let's fix up that link so that the user
navigates to the top level of the Controller API docs instead (as this
link did up until 4.8).

This closes https://github.com/ember-learn/guides-source/issues/1880.