Update quick-start.md
Some notes on my changes, as a complete and utter noob
Almost every documentation I see nowadays [or startup guide] neglects to mention how anything connects, making it almost impossible for someone that isn't introduced via another method to get started. Doing the steps and learning this information is always very simple, so there's no reason for not hand-holding first time users [and therefore making it more complicated than it actually is].

In the first change, my goal was to more clearly define what "outlet" is, and what the route is. The sentence "...any nested route will be rendered in that place" on line 80 is given before a route is defined, and it certainly doesn't define an outlet. It's like me telling you a Pompadigdog is actually a large grey Hoosmacker. If you don't know either of those terms, because i made them up, then it's difficult to understand the relation. It seems like these are also undefined terms that y'all made up, so... :[

Later I make a change to the explanation for the example on line 135 because I like understanding why I'm putting in all these non-focused pieces of code even if they're not part of the subject of discussion. Namely, the ".extend". In the way I read this, knowing very little, this quickstart guide tells me that a route is a page I'm loading, and that a >>page<< is returning data. However, I'm not calling anything at this point so what ends up happening is that I start asking questions about what the model actually is. The first thought is that maybe I don't understand it. So my option is to look at the example again, and see what about it is actually returning data. And now I encounter >>another<< new thing that I don't understand, and so my only option is believing that the other unknown has a hand in my confusion. Because now I can't conclusively solve this puzzle, I'm stuck looking at the rest of the tutorial for clues, or giving up.  While it may not be the focus of the guide, we should absolutely explain every piece so that the confusion is absolutely minimal, and if a user doesn't understand something, the effect will be minimized and they will still be able to continue the tutorial.

Please reword anything and everything I wrote to fit better into your work.